+++ source/MiniViewsLib/MainDialog/AdvancedTab.cpp
@@ -22,7 +22,7 @@ AdvancedWindow::AdvancedWindow() : blockEditNotify(true), lastActionClearedSetti
 
 }
 
-bool AdvancedWindow::CreateThis(HWND hParent, u32 windowId)
+bool AdvancedWindow::CreateThis(HWND hParent, u64 windowId)
 {
 	if ( getHandle() != NULL )
 		return SetParent(hParent);
